Video 20. Adding required functionality to deleting a customer.

+++ src/main/java/guru/springframework/services/reposervices/CustomerServiceRepoImpl.java
@@ -3,11 +3,13 @@ package guru.springframework.services.reposervices;
 import guru.springframework.commands.CustomerForm;
 import guru.springframework.converters.CustomerFormToCustomer;
 import guru.springframework.domain.Customer;
+import guru.springframework.exceptions.DBFailureException;
 import guru.springframework.exceptions.NotFoundException;
 import guru.springframework.repositories.CustomerRepository;
 import guru.springframework.services.CustomerService;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -50,7 +52,7 @@ public class CustomerServiceRepoImpl implements CustomerService {
     public Customer saveOrUpdateCustomerForm(CustomerForm customerForm) {
         Customer newCustomer = customerFormToCustomer.convert(customerForm);
 
-        if(newCustomer.getUser().getId() != null){
+        if (newCustomer.getUser().getId() != null) {
             Customer existingCustomer = getById(newCustomer.getId());
 
             newCustomer.getUser().setEnabled(existingCustomer.getUser().getEnabled());
@@ -60,7 +62,12 @@ public class CustomerServiceRepoImpl implements CustomerService {
     }
 
     @Override
+    @Transactional
     public void delete(Integer id) {
+        Optional<Customer> optionalCustomer = customerRepository.findById(id);
+        if (!optionalCustomer.isPresent()) {
+            throw new DBFailureException("unable to delete customer with " + id);
+        }
         customerRepository.deleteById(id);
     }
 }
